refactor: move product service into services folder
