+++ src/modules/product/product.controller.ts
@@ -12,7 +12,7 @@ import { ERouteName } from 'src/common/enums/route-name.enum';
 import { ProductEntity } from 'src/entities/product.entity';
 import { ProductsFilterDto } from 'src/modules/product/dto/filter-products.dto';
 // import { JwtAuthGuard } from 'src/modules/auth/guards/jwtAuthGuard';
-import { ProductService } from 'src/modules/product/product.service';
+import { ProductService } from 'src/modules/product/services/product.service';
 
 @ApiTags('Product')
 @Controller(ERouteName.PRODUCTS_ROUTE)
