+++ src/modules/product/product.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { ProductController } from 'src/modules/product/product.controller';
-import { ProductService } from 'src/modules/product/product.service';
+import { ProductService } from 'src/modules/product/services/product.service';
 
 @Module({
   imports: [],
